# Reduce memory allocation in arrowutils.SortRecord