This is my last pass before I move on to introduce the changes in `parca`

Benchmark result

```
goos: darwin
goarch: amd64
pkg: github.com/polarsignals/frostdb/pqarrow/arrowutils
cpu: Intel(R) Core(TM) i7-7820HQ CPU @ 2.90GHz
             │   old.txt    │           new.txt            │
             │    sec/op    │   sec/op     vs base         │
SortRecord-8   7.492µ ± 10%   6.972µ ± 7%  ~ (p=0.093 n=6)

             │  old.txt   │              new.txt              │
             │    B/op    │    B/op     vs base               │
SortRecord-8   808.0 ± 0%   400.0 ± 0%  -50.50% (p=0.002 n=6)

             │   old.txt   │              new.txt              │
             │  allocs/op  │ allocs/op   vs base               │
SortRecord-8   15.000 ± 0%   8.000 ± 0%  -46.67% (p=0.002 n=6)
```

I really missed `errdefer` in `zig` manual release during `error` returns in annoying.